feat: 마이페이지 최적화 및 yarn 지원 복원
