+++ package.json
@@ -26,7 +26,6 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.21.0",
-    "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/postcss": "^4.1.3",
     "@types/node": "^22.13.11",
     "@types/react": "^19.0.12",
@@ -48,5 +47,6 @@
     "typescript": "4.9.5",
     "typescript-eslint": "^8.24.1",
     "vite": "^6.2.2"
-  }
+  },
+  "packageManager": "yarn@4.6.0+sha512.5383cc12567a95f1d668fbe762dfe0075c595b4bfff433be478dbbe24e05251a8e8c3eb992a986667c1d53b6c3a9c85b8398c35a960587fbd9fa3a0915406728"
 }
